[#30] Fix: 답글 달기 누르면 같은 단계 댓글들 전체가 입력폼 나오는 버그 수정
